LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/css/selectors/user-valid-user-invalid-multifield-inputs.tentative-expected.txt



FAIL <input type=date> keyboard behavior for :user-valid/:user-invalid. assert_equals: Date input value should not be set after partially inputting the date. expected "" but got "1"
FAIL <input type=time> keyboard behavior for :user-valid/:user-invalid. assert_equals: Time input value should not be set after partially inputting the time. expected "" but got "1"
FAIL <input type=datetime-local> keyboard behavior for :user-valid/:user-invalid. assert_equals: Datetime input value should not be set after partially inputting the dateTimeLocal. expected "" but got "1"
